Refactor event handling and geometry modules
- Removed the legacy `event_type.rs` file and replaced it with more granular event types including `FocusEvent`, `GestureEvent`, `KeyboardEvent`, `MediaEvent`, `PointerEvent`, `SystemEvent`, `TextInputEvent`, `TouchEvent`, `ValueChangedEvent`, `WheelEvent`, and `WindowEvent`.
- Introduced a new `Bounds` struct to manage UI element dimensions, including margin and padding.
- Added `Border` and `Offset` structs to represent border widths and geometric offsets respectively.
- Updated `Point` and `Rectangle` structs to use `f32` for coordinates and dimensions instead of `i64`.
- Removed deprecated input event handling code and related modules.
- Enhanced the derive macro for events to require a category attribute for better event categorization.

+++ engage-ux-core/src/modules/geometry/point.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
 pub struct Point {
 	/// The X coordinate
-	pub x: i64,
+	pub x: f32,
 	/// The Y coordinate
-	pub y: i64,
+	pub y: f32,
 }
